Remove data manipulation logic from main plugin file
ISSUE PLSW-4
