+++ Packlink/Packlink.php
@@ -3,24 +3,16 @@
 namespace Packlink;
 
 require_once __DIR__ . '/vendor/autoload.php';
+require_once __DIR__ . '/CSRFWhitelistAware.php';
 
 use Doctrine\ORM\EntityManager;
 use Logeecom\Infrastructure\Logger\Logger;
-use Logeecom\Infrastructure\ORM\RepositoryRegistry;
-use Logeecom\Infrastructure\ServiceRegister;
 use Packlink\Bootstrap\Bootstrap;
 use Packlink\Bootstrap\Database;
-use Packlink\BusinessLogic\Configuration;
-use Packlink\BusinessLogic\ShippingMethod\Utility\ShipmentStatus;
-use Packlink\Entities\ShippingMethodMap;
 use Shopware\Components\Plugin;
 use Shopware\Components\Plugin\Context\InstallContext;
 use Shopware\Components\Plugin\Context\UninstallContext;
-use Shopware\Models\Dispatch\Dispatch;
-use Shopware\Models\Order\Status;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-
-require_once __DIR__ . '/CSRFWhitelistAware.php';
 
 class Packlink extends Plugin
 {
@@ -54,9 +46,6 @@ class Packlink extends Plugin
         $db = new Database($entityManager);
         $db->install();
         Logger::logInfo("Database for version [{$context->getCurrentVersion()}] created...", 'Integration');
-
-        $this->setDefaultConfig();
-
         Logger::logInfo('Installation completed...', 'Integration');
     }
 
@@ -65,7 +54,6 @@ class Packlink extends Plugin
      *
      * @param \Shopware\Components\Plugin\Context\UninstallContext $context
      *
-     * @throws \Doctrine\ORM\OptimisticLockException
      * @throws \Logeecom\Infrastructure\ORM\Exceptions\RepositoryNotRegisteredException
      */
     public function uninstall(UninstallContext $context)
@@ -73,85 +61,10 @@ class Packlink extends Plugin
         Bootstrap::init();
 
         if (!$context->keepUserData()) {
-            $this->removeCreatedShippingMethods();
-
             /** @var EntityManager $entityManager */
             $entityManager = $this->container->get('models');
             $db = new Database($entityManager);
             $db->uninstall();
         }
-    }
-
-    /**
-     * Sets default configuration.
-     *
-     * @throws \Logeecom\Infrastructure\TaskExecution\Exceptions\TaskRunnerStatusStorageUnavailableException
-     */
-    protected function setDefaultConfig()
-    {
-        $this->setDefaultStatusMapping();
-        $this->getConfigService()->setTaskRunnerStatus('', null);
-    }
-
-    /**
-     * Sets default order status mapping.
-     */
-    protected function setDefaultStatusMapping()
-    {
-        $mappings = $this->getConfigService()->getOrderStatusMappings();
-
-        if (empty($mappings)) {
-            $this->getConfigService()->setOrderStatusMappings(
-                [
-                    ShipmentStatus::STATUS_PENDING => '',
-                    ShipmentStatus::STATUS_ACCEPTED => Status::ORDER_STATE_READY_FOR_DELIVERY,
-                    ShipmentStatus::STATUS_READY => Status::ORDER_STATE_READY_FOR_DELIVERY,
-                    ShipmentStatus::STATUS_IN_TRANSIT => Status::ORDER_STATE_READY_FOR_DELIVERY,
-                    ShipmentStatus::STATUS_DELIVERED => Status::ORDER_STATE_COMPLETELY_DELIVERED,
-                ]
-            );
-        }
-    }
-
-    /**
-     * Retrieves configuration service;
-     *
-     * @return \Packlink\Services\BusinessLogic\ConfigurationService
-     */
-    protected function getConfigService()
-    {
-        if ($this->configService === null) {
-            $this->configService = ServiceRegister::getService(Configuration::CLASS_NAME);
-        }
-
-        return $this->configService;
-    }
-
-    /**
-     * Removes creates shipping methods during install.
-     *
-     * @throws \Logeecom\Infrastructure\ORM\Exceptions\RepositoryNotRegisteredException
-     * @throws \Doctrine\ORM\OptimisticLockException
-     */
-    protected function removeCreatedShippingMethods()
-    {
-        $dispatchRepository = Shopware()->Models()->getRepository(Dispatch::class);
-
-        $mapRepository = RepositoryRegistry::getRepository(ShippingMethodMap::getClassName());
-        $maps = $mapRepository->select();
-
-        /** @var ShippingMethodMap $map */
-        foreach ($maps as $map) {
-            if ($dispatch = $dispatchRepository->find($map->shopwareCarrierId)) {
-                Shopware()->Models()->remove($dispatch);
-            }
-        }
-
-        $backupId = $this->getConfigService()->getBackupCarrierId();
-        if ($backupId && $dispatch = $dispatchRepository->find($backupId)) {
-            Shopware()->Models()->remove($dispatch);
-        }
-
-        Shopware()->Models()->flush();
     }
 }
